requirements.txt
# PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Models
timm>=0.9.12

# Data processing
numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.4.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.7.0

# Dimensionality reduction
umap-learn>=0.5.3

# Progress bar
tqdm>=4.65.0

# Config
pyyaml>=6.0

# Download dataset
gdown>=4.7.1

# Mixed precision
accelerate>=0.25.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
